+++ adapter/spark/src/main/java/com/asml/apa/wta/spark/streams/TaskMetrics.java
@@ -1,0 +1,14 @@
+package com.asml.apa.wta.spark.streams;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+/**
+ * Task metrics record.
+ *
+ * @author Atour Mousavi Gourabi
+ * @since 1.0.0
+ */
+@Getter
+@AllArgsConstructor
+public class TaskMetrics {}