+++ core/src/main/java/com/asml/apa/wta/core/model/Task.java
@@ -1,7 +1,9 @@
 package com.asml.apa.wta.core.model;
 
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Data;
+import lombok.Getter;
 import lombok.Setter;
 import org.apache.avro.Schema;
 import org.apache.avro.SchemaBuilder;
@@ -19,8 +21,8 @@ import org.apache.avro.generic.GenericRecord;
 public class Task implements BaseTraceObject {
   private static final long serialVersionUID = -1372345471722101373L;
 
-  @Setter
-  private static String schemaVersion;
+  @Getter(value = AccessLevel.NONE)
+  private final String schemaVersion = this.getSchemaVersion();
 
   private final long id;