+++ core/src/main/java/com/asml/apa/wta/core/model/Workload.java
@@ -2,8 +2,10 @@ package com.asml.apa.wta.core.model;
 
 import com.asml.apa.wta.core.model.enums.Domain;
 import java.util.Arrays;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Data;
+import lombok.Getter;
 import lombok.Setter;
 import org.apache.avro.Schema;
 import org.apache.avro.SchemaBuilder;
@@ -21,8 +23,8 @@ import org.apache.avro.generic.GenericRecord;
 public class Workload implements BaseTraceObject {
   private static final long serialVersionUID = -4547341610378381743L;
 
-  @Setter
-  private static String schemaVersion;
+  @Getter(value = AccessLevel.NONE)
+  private final String schemaVersion = this.getSchemaVersion();
 
   private final Workflow[] workflows;